# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a server ecommerce.",
        "title": "Swagger Golang Ecommerce",
        "contact": {
            "name": "Nguyễn Đại Nghĩa",
            "url": "https://www.facebook.com/shy.leopard.beo",
            "email": "nghiabeo1605@gmail.com"
        }
    },
    "host": "localhost:8080",
    "paths": {
...
        "/api/product-types": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get all ProductType",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ProductType"
                ],
                "summary": "Get all ProductType",
                "parameters": [
                    {
                        "type": "integer",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "order",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "search",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "array",
                                "items": {
                                    "$ref": "#/definitions/github_com_NghiaLeopard_Go-Ecommerce-Backend_internal_api_handler_response.ProductType"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create ProductType",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ProductType"
                ],
                "summary": "Create ProductType",
                "parameters": [
                    {
                        "description": "Create ProductType",
                        "name": "tags",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_NghiaLeopard_Go-Ecommerce-Backend_internal_api_handler_request.CreateProductType"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_NghiaLeopard_Go-Ecommerce-Backend_internal_api_handler_response.ProductType"
                        }
                    }
                }
            }
        },
        "/api/product-types/delete-many": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Delete many ProductType",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ProductType"
                ],
                "summary": "Delete many ProductType",
                "parameters": [
                    {
                        "description": "DeleteMany ProductType",
                        "name": "tags",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_NghiaLeopard_Go-Ecommerce-Backend_internal_api_handler_request.DeleteManyProductType"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Delete many ProductType success",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/product-types/{ProductTypeId}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get ProductType by id",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ProductType"
                ],
                "summary": "Get ProductType by id",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "User ID",
                        "name": "ProductTypeId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_NghiaLeopard_Go-Ecommerce-Backend_internal_api_handler_response.ProductType"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Update ProductType",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ProductType"
                ],
                "summary": "Update ProductType",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Update ProductType",
                        "name": "ProductTypeId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Update ProductType",
                        "name": "tags",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_NghiaLeopard_Go-Ecommerce-Backend_internal_api_handler_request.GetBodyUpdateProductType"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_NghiaLeopard_Go-Ecommerce-Backend_internal_api_handler_response.ProductType"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Delete ProductType",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ProductType"
                ],
                "summary": "Delete ProductType",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Delete ProductType",
                        "name": "ProductTypeId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "github_com_NghiaLeopard_Go-Ecommerce-Backend_internal_api_handler_request.CreateProductType": {
            "type": "object",
            "required": [
                "name",
                "slug"
            ],
            "properties": {
                "name": {
                    "type": "string",
                    "minLength": 1
                },
                "slug": {
                    "type": "string",
                    "minLength": 1
                }
            }
        },
...
        "github_com_NghiaLeopard_Go-Ecommerce-Backend_internal_api_handler_request.DeleteManyProductType": {
            "type": "object",
            "required": [
                "arrayId"
            ],
            "properties": {
                "arrayId": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                }
            }
        },
...
        "github_com_NghiaLeopard_Go-Ecommerce-Backend_internal_api_handler_request.GetBodyUpdateProductType": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "slug": {
                    "type": "string"
                }
            }
        },
...
        "github_com_NghiaLeopard_Go-Ecommerce-Backend_internal_api_handler_response.ProductType": {
            "type": "object",
            "properties": {
                "_id": {
                    "type": "integer"
                },
                "create_at": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "slug": {
                    "type": "string"
                },
                "update_at": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Add prefix of Bearer before  token Ex: \"Bearer token\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
...
}